extract JWT authentication to extension and add config validation
